requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai
